Extract venue Cypher show query productions section
