+++ src/neo4j/cypher-queries/venue/show/index.js
@@ -1,0 +1,7 @@
+import getShowQuery from './show';
+import getShowProductionsQuery from './show-productions';
+
+export default () => [
+	getShowQuery(),
+	getShowProductionsQuery()
+];
